+++ Cookie/Core/Character.cs
@@ -81,7 +81,7 @@ namespace Cookie.Core
             account.Network.RegisterPacket<CharacterSelectedSuccessMessage>(HandleCharacterSelectedSuccessMessage,
                 MessagePriority.VeryHigh);
 
-            #endregion
+            #endregion Choice Handler
 
             #region Creation Handler
 
@@ -94,14 +94,14 @@ namespace Cookie.Core
             account.Network.RegisterPacket<CharacterNameSuggestionSuccessMessage>(
                 HandleCharacterNameSuggestionSuccessMessage, MessagePriority.VeryHigh);
 
-            #endregion
+            #endregion Creation Handler
 
             #region Deletion Handler
 
             account.Network.RegisterPacket<CharacterDeletionErrorMessage>(HandleCharacterDeletionErrorMessage,
                 MessagePriority.VeryHigh);
 
-            #endregion
+            #endregion Deletion Handler
 
             #region Stats Handler
 
@@ -118,7 +118,7 @@ namespace Cookie.Core
             account.Network.RegisterPacket<UpdateLifePointsMessage>(HandleUpdateLifePointsMessage,
                 MessagePriority.VeryHigh);
 
-            #endregion
+            #endregion Stats Handler
 
             #region Initialization Handler
 
@@ -129,7 +129,7 @@ namespace Cookie.Core
             account.Network.RegisterPacket<SetCharacterRestrictionsMessage>(HandleSetCharacterRestrictionsMessage,
                 MessagePriority.VeryHigh);
 
-            #endregion
+            #endregion Initialization Handler
 
             _account.Network.RegisterPacket<MapComplementaryInformationsDataMessage>(
                 HandleMapComplementaryInformationsDataMessage, MessagePriority.Normal);
@@ -143,7 +143,7 @@ namespace Cookie.Core
 
         public CharacterStatus Status { get; set; }
 
-        public ulong Id { get; set; }
+        public double Id { get; set; }
         public string Name { get; set; }
         public int Level { get; set; }
         public bool Sex { get; set; }
@@ -152,10 +152,10 @@ namespace Cookie.Core
 
         public BreedEnum Breed { get; set; }
 
-        public int LifePercentage => (int) (Stats.LifePoints / (double) Stats.MaxLifePoints * 100);
-        public int WeightPercentage => (int) (Weight / (double) MaxWeight * 100);
-        public int EnergyPercentage => (int) (Stats.EnergyPoints / (double) Stats.MaxEnergyPoints * 100);
-        public int ExperiencePercentage => (int) (Stats.Experience / (double) Stats.ExperienceNextLevelFloor * 100);
+        public int LifePercentage => (int)(Stats.LifePoints / (double)Stats.MaxLifePoints * 100);
+        public int WeightPercentage => (int)(Weight / (double)MaxWeight * 100);
+        public int EnergyPercentage => (int)(Stats.EnergyPoints / (double)Stats.MaxEnergyPoints * 100);
+        public int ExperiencePercentage => (int)(Stats.Experience / (double)Stats.ExperienceNextLevelFloor * 100);
 
         public int CellId { get; set; }
         public int MapId { get; set; }
@@ -302,7 +302,7 @@ namespace Cookie.Core
             Logger.Default.Log("Une erreur est survenue lors de la suppresion du personnae.");
         }
 
-        #endregion
+        #endregion Deletion
 
         private void HandleMapComplementaryInformationsDataMessage(IAccount account,
             MapComplementaryInformationsDataMessage message)
@@ -371,10 +371,10 @@ namespace Cookie.Core
             account.Character.Name = message.Infos.Name;
             account.Character.Sex = message.Infos.Sex;
             account.Character.Look = message.Infos.EntityLook;
-            account.Character.Breed = (BreedEnum) message.Infos.Breed;
+            account.Character.Breed = (BreedEnum)message.Infos.Breed;
         }
 
-        #endregion
+        #endregion Choice
 
         #region Creation
 
@@ -383,12 +383,12 @@ namespace Cookie.Core
         {
             // Si nous ne pouvons pas créer de personnages, nous arrêtons la fonction
             if (!message.YesYouCan) return;
-            // Sinon, nous choisissons une classe au hasard 
-            var breedId = (byte) Randomize.GetRandomNumber(1, 18);
+            // Sinon, nous choisissons une classe au hasard
+            var breedId = (byte)Randomize.GetRandomNumber(1, 18);
             // Nous récupérons les informations de la classe avec les D2O
             var breed = ObjectDataManager.Instance.Get<Breed>(breedId);
             // Nous récupérons la couleur de base de la classe, et nous faisons un léger random sur la couleur
-            var breedColors = breed.MaleColors.Select(i => Randomize.GetRandomNumber((int) i - 80000, (int) i + 80000))
+            var breedColors = breed.MaleColors.Select(i => Randomize.GetRandomNumber((int)i - 80000, (int)i + 80000))
                 .ToList();
             // On récupère la liste des cosmetics disponibles pour cette classe et ce sexe
             var headsList = ObjectDataManager.Instance.EnumerateObjects<Head>().ToList()
@@ -397,35 +397,43 @@ namespace Cookie.Core
             var head = headsList[Randomize.GetRandomNumber(0, 7)];
             //// Nous envoyons la requête pour créer le personnage
             var test = new CharacterCreationRequestMessage(account.Character.Name, breedId, false, breedColors,
-                (ushort) head.Id);
+                (ushort)head.Id);
             account.Network.SendToServer(test);
         }
 
         private void HandleCharacterCreationResultMessage(IAccount account, CharacterCreationResultMessage message)
         {
-            switch ((CharacterCreationResultEnum) message.Result)
+            switch ((CharacterCreationResultEnum)message.Result)
             {
                 case CharacterCreationResultEnum.OK:
                     account.Character.IsFirstConnection = true;
                     break;
+
                 case CharacterCreationResultEnum.ERR_NO_REASON:
                     break;
+
                 case CharacterCreationResultEnum.ERR_INVALID_NAME:
                     Logger.Default.Log("Ce nom de personnage est invalide.", LogMessageType.Public);
                     break;
+
                 case CharacterCreationResultEnum.ERR_NAME_ALREADY_EXISTS:
                     Logger.Default.Log("Ce nom de personnage est déjà pris.", LogMessageType.Public);
                     break;
+
                 case CharacterCreationResultEnum.ERR_TOO_MANY_CHARACTERS:
                     Logger.Default.Log("Vous avez déjà atteint la limite de personnages disponible.",
                         LogMessageType.Public);
                     break;
+
                 case CharacterCreationResultEnum.ERR_NOT_ALLOWED:
                     break;
+
                 case CharacterCreationResultEnum.ERR_NEW_PLAYER_NOT_ALLOWED:
                     break;
+
                 case CharacterCreationResultEnum.ERR_RESTRICED_ZONE:
                     break;
+
                 case CharacterCreationResultEnum.ERR_INCONSISTENT_COMMUNITY:
                     break;
             }
@@ -448,7 +456,7 @@ namespace Cookie.Core
             account.Network.SendToServer(test);
         }
 
-        #endregion
+        #endregion Creation
 
         #region Stats
 
@@ -500,7 +508,7 @@ namespace Cookie.Core
             account.Character.Stats.MaxLifePoints = message.MaxLifePoints;
         }
 
-        #endregion
+        #endregion Stats
 
         #region Initialization
 
@@ -525,6 +533,6 @@ namespace Cookie.Core
             account.Character.Restrictions = message.Restrictions;
         }
 
-        #endregion
+        #endregion Initialization
     }
 }
