+++ Cookie/Game/Entity/Player.cs
@@ -4,7 +4,7 @@ namespace Cookie.Game.Entity
 {
     public class Player : IPlayer, IEntity
     {
-        public Player(int cell, int id, string name)
+        public Player(int cell, double id, string name)
         {
             CellId = cell;
             Id = id;
@@ -13,7 +13,7 @@ namespace Cookie.Game.Entity
 
         public int CellId { get; set; }
 
-        public int Id { get; set; }
+        public double Id { get; set; }
 
         public string Name { get; set; }
     }
